Restructure MiniMax GGUF runtime into generic loader/components
Move generic GGUF runtime out of the MiniMax model files per the
loader/components boundary:

- src/loader/gguf/: file loading, block metadata, row readers, data
  structures (quant_types, quantized_tensor, quantized_loader)
- src/components/gguf/: CUDA kernel-facing op wrappers + TP logits
  helpers (quantized_ops, tp_logits)
- src/models/minimax_m2/: MiniMax-specific architecture/assembly/CLI
  glue only (runtime->architecture, loader->gguf_model,
  generation->gguf_generation; cli generate_minimax->generate_minimax_gguf)

Behavior unchanged: TP4 smoke still yields generated_token_ids=16474,9
at 15.853 GiB/rank.
