+++ src/cli/generate_minimax_gguf.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from src.models.minimax_m2.generation import main
+from src.models.minimax_m2.gguf_generation import main
 
 
 if __name__ == "__main__":
