+++ src/components/gguf/__init__.py
@@ -1,0 +1,1 @@
+"""Kernel-facing GGUF runtime components."""
