+++ tests/test_gguf_quantized_ops_cuda.py
@@ -5,10 +5,9 @@ from pathlib import Path
 import pytest
 import torch
 
+from src.kernels.cuda_loader import load_cuda_kernel
 from src.loader.gguf.bundle import read_gguf_bundle
 from src.loader.gguf.tensor_reader import GGUFTensorDataReader
-from src.kernels.cuda_loader import load_cuda_kernel
-from src.models.minimax_m2.loader import MiniMaxM2GGUFLoader
 
 
 REAL_MINIMAX_PATH = Path("/mnt/data1/dsv4_inference/gguf_hfd/MiniMax-M2.7-GGUF/UD-IQ1_M")
@@ -71,6 +70,7 @@ def test_real_minimax_q4_q5_gemm_matches_reference(name: str, type_id: int) -> N
     assert float((y - expected).abs().max().item()) < 2.0e-2
     assert float((yp - expected).abs().max().item()) < 2.0e-2
 
+
 @pytest.mark.skipif(not (REAL_MINIMAX_PATH.exists() and _cuda_q4_q5_available()), reason="real MiniMax GGUF or CUDA q4/q5 extension not available")
 def test_real_minimax_q5_pair_gemm_matches_separate_outputs() -> None:
     bundle = read_gguf_bundle(REAL_MINIMAX_PATH)
@@ -90,58 +90,3 @@ def test_real_minimax_q5_pair_gemm_matches_separate_outputs() -> None:
     assert bool(torch.isfinite(pv).all().item())
     assert float((pk.float() - yk).abs().max().item()) == 0.0
     assert float((pv.float() - yv).abs().max().item()) == 0.0
-
-
-@pytest.mark.skipif(not (REAL_MINIMAX_PATH.exists() and _cuda_q4_q5_available()), reason="real MiniMax GGUF or CUDA q4/q5 extension not available")
-def test_minimax_loader_keeps_dense_weights_as_raw_blocks() -> None:
-    loader = MiniMaxM2GGUFLoader(
-        REAL_MINIMAX_PATH,
-        device="cuda:0",
-        n_layers=1,
-        expert_start=0,
-        expert_count=1,
-        preload_moe=False,
-    )
-    try:
-        model = loader.load()
-    finally:
-        loader.close()
-    assert model.embedding.tensor.blocks.dtype == torch.uint8
-    assert model.embedding.tensor.type_name == "q4_k"
-    assert model.lm_head.tensor.type_name == "q4_k"
-    assert model.layers[0].attention.q_proj.tensor.type_name == "q5_k"
-    assert model.layers[0].attention.k_proj.tensor.type_name == "q5_k"
-    assert model.layers[0].attention.v_proj.tensor.type_name == "q5_k"
-    assert model.layers[0].attention.o_proj.tensor.type_name == "q5_k"
-    assert tuple(model.embedding.tensor.blocks.shape[1:]) == (12, 144)
-    assert tuple(model.layers[0].attention.q_proj.tensor.blocks.shape[1:]) == (12, 176)
-
-
-@pytest.mark.skipif(not (REAL_MINIMAX_PATH.exists() and _cuda_q4_q5_available()), reason="real MiniMax GGUF or CUDA q4/q5 extension not available")
-def test_minimax_one_layer_forward_tp1_smoke() -> None:
-    loader = MiniMaxM2GGUFLoader(
-        REAL_MINIMAX_PATH,
-        device="cuda:0",
-        n_layers=1,
-        expert_start=0,
-        expert_count=256,
-        preload_moe=True,
-    )
-    try:
-        model = loader.load()
-    finally:
-        loader.close()
-    model.reset_cache(batch_size=1, max_seq_len=4)
-    tokens = torch.tensor([[1, 2, 3]], device="cuda:0", dtype=torch.long)
-    next_token = model.forward(tokens, 0, return_next_token=True)
-    assert next_token.shape == (1,)
-    assert bool(torch.isfinite(next_token.float()).all().item())
-
-
-def test_minimax_loader_rank_row_range_splits_vocab_rows() -> None:
-    assert MiniMaxM2GGUFLoader._rank_row_range(16, 4, 0) == (0, 4)
-    assert MiniMaxM2GGUFLoader._rank_row_range(16, 4, 3) == (12, 4)
-    assert MiniMaxM2GGUFLoader._rank_row_range(10, 4, 0) == (0, 2)
-    assert MiniMaxM2GGUFLoader._rank_row_range(10, 4, 1) == (2, 3)
-    assert MiniMaxM2GGUFLoader._rank_row_range(10, 4, 2) == (5, 2)
-    assert MiniMaxM2GGUFLoader._rank_row_range(10, 4, 3) == (7, 3)
